requirements.txt
alembic==1.9.3
aniso8601==9.0.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-editor==1.0.4
pytz==2022.7.1
rauth==0.7.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.13.0
flask_socketio==5.3.2
numpy~=1.24.2
gevent~=22.10.2
PyYAML~=6.0